refactor
extra files for class ResourceManager and FileIO.
comment out __checkLine() to test until generating machinecode works
