--- py-assembler/output.o
@@ -1,3 +1,0 @@
-000010000011010001100000110000100000100100LW00001000010
-SW00001000010
-00011000001000100000000000000000
